+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to the "Range Slider Panel" plugin will be documented in thi
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.1] - 2024-01-18
+- Add default prefix, delimiter and suffix if it is not defined
+- Allow select of variable from options
+
 ## [1.0.0] - 2024-01-08
 
 - Initial release of the Range Slider Panel plugin.